Update HDF5 libraries (Clean .h files, rebuild with multithreaded DLL option). Removed .project, .setting files. Moved all C# related files to CSharpAPI. Cleaned EosPack library from compilation warnings. Add unit tests for PTDiagramCalculator.
work-items: #14404, #14406, #14407, #14409
git-tfs-id: [https://sede-pt-ce2.visualstudio.com/DefaultCollection]$/Basin Modeling/IBS/Trunk;C34705

+++ development/libraries/EosPack/src/EosPvtModel.C
@@ -1093,7 +1093,7 @@ int EosPvtModel::SetPointers( void )
    m_pEosPvtTable->SetAbcPointers( iN, &m_pAbcOffset, &pFinal );
 
    /* Calculate the total length */
-   i = pFinal - m_pWorkArray;
+   i = (int)(pFinal - m_pWorkArray);
    i = ( i > i1 ? i : i1 );
 
    /* End of the routine */
